+++ main/Sequencer/Message.h
@@ -65,13 +65,13 @@ public:
 
 	// This message to send simple messages, with no response
 	static Message *future_Message(TASK_NAME target, TASK_NAME from,
-			int _event, int val, int rate);
+			int _event, long int val, long int rate);
 
 	int event;            // A 'valid' event for the device. If the device receives
 	                      // an invalid event, ignore it. If this is a response, this
 	                      // is the message type of the requesting message.
-	int value;            //  The value we want to set.
-	int rate;             // An indication of how fast this should happen.
+	long int value;            //  The value we want to set.
+	long int rate;             // An indication of how fast this should happen.
 
 protected:
 	Message ();
